Workaround for iOS bug
More information about the bug here: https://bugs.webkit.org/show_bug.cgi?id=176896. The bug has supposedly been fixed and should be in the next release 11.3. As of the time of this change the current version is 11.2. Once the bug has been confirmed to be fixed we can remove this workaround.