Fix for condition in some browsers where onchange was not firing.
In some browsers (for example Safari on iOS) the onchange event is not fired after explicitly setting its value. Because we are setting the input element's value during the oninput event in order to apply formatting this means we need to check if the value changed during the onblur event.