Redesign of apex component.
The apex component had to be completely redesigned so that it no longer takes care of enqueuing the action. Calling component.get('c.actionName') was inconsistent and often resulted in an 'invalid key' error. I don't know exactlly what caused the problem, but rather than spend a lot of time troubleshooting only to end up likely needing to change the component, I decided to just redesign it to not enqueue the action. The consumer of the component is now responsible for enqueuing the action and the apex component simply provides the handler function passed to action.setCallback().