Another redesign of the apex component.
This design is more close to the original design in that the component can now take care of configuring and enqueuing the action. This was achieved by using a module pattern similar to the utils component. Instead of exposing the execute() method as an aura:method we have a getModule() aura:method which returns an object that has an execute() method. Because the execute() method is no longer an aura:method the framework does not interfere with the arguments passed to execute().