Do not prevent entering a period if scale is 0.
Allow users to type a period even if the scale attribute is 0. This allows for entry of values such as "9.5k" which will be expanded to "9,500." If a user enters a decimal number (e.g., "2.5") the period will be removed when the value is rounded and reformatted in the onblur or onchange event handlers.