Added support for important fields in inputLookup component.
You can now prefix a field with a "!" character to indicate that the field is important and should appear in the search results even if the user does not normally have access to that field. Additionally, a fix was made so that the ID and Name fields always are included.