Added ability to specify S-Object display order.
A new displayOrder property was added to allow for explicit control over the display order of the S-Objects in the object switcher. This new property is optional. S-Objects with no displayOrder will appear after S-Objects with a displayOrder in alphabetical order. Two or more S-Objects with the same displayOrder will be sorted alphabetically.